requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
numpy>=1.26.0
python-multipart==0.0.6
pydantic==1.10.13
pillow==10.1.0
python-dotenv==1.0.0
matplotlib>=3.8.0
langchain>=0.0.350
langchain-community>=0.0.13
langchain-core>=0.1.23
rasterio>=1.3.0
geopandas>=0.14.0
requests>=2.31.0 